Changed to MIT license for reasons:
At this point my plan is to blatantly copy the function names and
partial behavior from Mocha, which is MIT. I'm not sure that
matters, but I think MIT is less likely to cuase problems later
down the road with this project. I could also see them copying
this in their project, which I don't mind. For that reason I also
think MIT is less likely to become a problem.