Refactored and using index.js as API into rest of system:
The goal is to make this package installable via npm, and then be
able to just require('living-tests'). I have not tested that yet,
but it might work now. index.js is an API into the rest of the
system, and this has been partially tested, because the test that
tests living-tests with living-tests is now using index.js in such
a way. But that is still slightly different from how someone else
would require it, so I need to test that as well.