Greatly simplified!:
I was able to delete the test class and the index.js file. Your
tests will be much simpler and clearer now. When you
require('living-tests') that will bring in an assert object, which
is the only thing you need to pull in from this package. Then you
just create a setup function locally and call it as needed before
test code and use that assert object to call assert.equal() for all
your assertions.