Deleted a bunch of commented out stuff:
This is VimL code from the elhiv library that I planned to
translate to javascript as I make the elhin library. But that is
taking longer and requiring more work than I would like, and it is
not fun so I am taking more of a clean slate approach now.