refactored:
renamed String.GetVimNoMagic() to String.getVimNoMagic(). This is
in harmony with project naming standards. Method names should start
with a lowercase letter. Normally camel casing is not allowed, but
in this case it is required. That is because it is a monkey patch
that adds a method to the String prototype. All the other String
methods are named that way, so this fits well with them. This will
also make it harder to ignore if eventually a method with the same
name gets added to the JavaScript String prototype.