Increased code coverage:
The if that was deleted made the test more safe to run. I think
relying on an if like that indicates some kind of problem with the
tests. I'm not sure the best way to address it yet, but for now I
think it is safe enough as is and 100% coverage is worth it.